Classe monstre + nom des paramètres

+++ HugoLand/Métier/Item.cs
@@ -16,14 +16,34 @@ namespace HugoLand.Métier
     {
         public static void CréerItem(string Description, ICollection<Accès_aux_données.EffetItem> effet, int imageid,
             ICollection<Accès_aux_données.InventaireHero> invHe, Accès_aux_données.Monde monde, int mondeid, int niveau, string nom, int poids, int quantite,
-            bool ramasse, int reqForce, int reqDex, int reqEnd, int reqInt, int reqNiv, byte[] rowVersion, int valeur)
+            bool ramasse, int reqForce, int reqDex, int reqEnd, int reqInt, int reqNiv, byte[] rowVersion, decimal valeur)
         {
             using (Accès_aux_données.Entities context = new Accès_aux_données.Entities())
             {
                 var items = context.Set<Accès_aux_données.Item>();
-                items.Add(new Accès_aux_données.Item() { x = 0, y = 0, Description = Description, EffetItems = effet, ImageId = imageid
-                , InventaireHeroes = invHe, Monde = monde, MondeId = mondeid, Niveau = niveau, Nom = nom, Poids = poids, Quantite = quantite, Ramasse = ramasse
-                , ReqDexterite = reqDex, ReqEndurance = reqEnd, ReqForce = reqForce, ReqIntelligence = reqInt, ReqNiveau = reqNiv, RowVersion = rowVersion, ValeurArgent = valeur});
+                items.Add(new Accès_aux_données.Item()
+                {
+                    x = 0,
+                    y = 0,
+                    Description = Description,
+                    EffetItems = effet,
+                    ImageId = imageid,
+                    InventaireHeroes = invHe,
+                    Monde = monde,
+                    MondeId = mondeid,
+                    Niveau = niveau,
+                    Nom = nom,
+                    Poids = poids,
+                    Quantite = quantite,
+                    Ramasse = ramasse,
+                    ReqDexterite = reqDex,
+                    ReqEndurance = reqEnd,
+                    ReqForce = reqForce,
+                    ReqIntelligence = reqInt,
+                    ReqNiveau = reqNiv,
+                    RowVersion = rowVersion,
+                    ValeurArgent = valeur
+                });
                 context.SaveChanges();
             }
         }
@@ -32,7 +52,8 @@ namespace HugoLand.Métier
         {
             using (Accès_aux_données.Entities context = new Accès_aux_données.Entities())
             {
-                context.Items.Remove(item);
+                if (item != null)
+                    context.Items.Remove(item);
                 context.SaveChanges();
             }
         }
@@ -41,7 +62,8 @@ namespace HugoLand.Métier
         {
             using (Accès_aux_données.Entities context = new Accès_aux_données.Entities())
             {
-                item.Quantite = quantite;
+                if (item != null)
+                    item.Quantite = quantite;
                 context.SaveChanges();
             }
         }
